+++ build.ps1
@@ -10,7 +10,7 @@
         The default value is '' (empty string).
 
     .PARAMETER BuildConfig
-        Not yet written.
+        Path to a file with configuration. Supported extensions : psd1, yaml, yml, json, jsonc
 
     .PARAMETER OutputDirectory
         Specifies the folder to build the artefact into. The default value is 'output'.
@@ -22,7 +22,7 @@
     .PARAMETER RequiredModulesDirectory
         Can be a path (relative to $PSScriptRoot or absolute) to tell Resolve-Dependency
         and PSDepend where to save the required modules. It is also possible to use
-        'CurrentUser' och 'AllUsers' to install missing dependencies. You can override
+        'CurrentUser' or 'AllUsers' to install missing dependencies. You can override
         the value for PSDepend in the Build.psd1 build manifest. The default value is
         'output/RequiredModules'.
 
@@ -49,13 +49,13 @@
         used in the DscResource.Test.build.ps1 tasks.
 
     .PARAMETER ResolveDependency
-        Not yet written.
+        Resolve missing dependencies.
 
     .PARAMETER BuildInfo
         The build info object from ModuleBuilder. Defaults to an empty hashtable.
 
     .PARAMETER AutoRestore
-        Not yet written.
+        Specifies to restore the required modules by running build.ps1 with ResolveDependency switch and empty task `noop`.
 
     .PARAMETER UseModuleFast
         Specifies to use ModuleFast instead of PowerShellGet to resolve dependencies
@@ -506,7 +506,7 @@ begin
             {
                 $paramValue = $MyInvocation.BoundParameters.Item($cmdParameter)
 
-                Write-Debug " adding  $cmdParameter :: $paramValue [from user-provided parameters to Build.ps1]"
+                Write-Debug -Message " adding  $cmdParameter :: $paramValue [from user-provided parameters to Build.ps1]"
 
                 $resolveDependencyParams.Add($cmdParameter, $paramValue)
             }
@@ -517,7 +517,7 @@ begin
 
                 if ($paramValue)
                 {
-                    Write-Debug " adding  $cmdParameter :: $paramValue [from default Build.ps1 variable]"
+                    Write-Debug -Message " adding  $cmdParameter :: $paramValue [from default Build.ps1 variable]"
 
                     $resolveDependencyParams.Add($cmdParameter, $paramValue)
                 }